act_shared_auth_ui: mixin_auth_redirect pop all pages upon event
If MixinAuthRedirectService detects a sign out, it pop all pages and
show SignIn page instead of simply pushing SignInPage over current page.

This is maybe not perfect but this appears a reasonable workaround
for routeManager concurrency issue between app trying to pop all
and this mixin trying to push over.

Refs: #19632